UserUpdateIdmResourceTest test is extracted
